// YYC3-H/AI/src/app/types.ts
export interface FileNode {
  name: string
  type: 'file' | 'folder'
  children?: FileNode[]
  ext?: string
}